Allowing usage of vanilla IC2 machines
Vanilla IC2 machines can be crafted from Gregtech's Automatic versions
if necessary.
As a side note, Gregtech's Automatic Machines can be upgraded in a
similar fashion to the vanilla IC2 machines, only this time by
right-clicking the machine with the desired upgrade. Doing this though
will permanently upgrade that specific machine, and will not allow you
to take out the upgrade.