Thermal Expansion balancing...
Cobble gen has been disabled, and recipes for machines have now been
made to be more difficult. If you already made some of these machines,
you are in luck as they will not be removed.

Also, if you did not understand, COBBLESTONE GENERATORS OF ANY KIND ARE
NOT ALLOWED ON THE SERVER. You have been warned.