Added ability to view ticks per second in-game
While already installed on the server, this mod will allow players to
view the ticks per second (tps) of their own worlds if they wish. In
order to use this tool, simply type '/btps' in the in-game chat to view.
For more info about the mod, google "Bechill TPS mod".