Added ExtraUtils (Initial Nerf)
Added back ExtraUtils, especially due to useful less laggy alternative
to pumping lava from the Nether.

Usage of this new pump will now be enforced when attempting to pump lava
from the nether. Usage of any other pump in the Nether will no longer be
allowed on the server. To use this pump, it must be placed on the level
just above the surface of the lava, and does not need to be above lava
to be able to operate. It will replace the lava block it pumps out with
a stone block to avoid lag-inducing lava flow calculations.