Disable annoying invisible railcraft blocks
Railcraft uses invisible blocks to track player movement, but causes
issues with many things as a result, as it is not considered an air
block. As this disable changes the ID of that specific block, the
railcraft config file must be updated in order to continue playing on
the server.